+++ docs/.docit/docit.config.js
@@ -24,6 +24,10 @@ const config = defineConfig({
           path: "/document/MDX",
         },
         {
+          title: "中文测试",
+          path: "/configure/中文标题测试",
+        },
+        {
           title: "Live Block",
           path: "/document/live-block",
         },