+++ src/node/plugins/index.ts
@@ -1,9 +1,9 @@
-import { Plugin, defineConfig, mergeConfig } from "vite";
+import { defineConfig, mergeConfig, PluginOption } from "vite";
 import fsx from "fs-extra";
 import path from "path";
 import react from "@vitejs/plugin-react";
 import { mdx, getCompilerOptions } from "./mdx/index.js";
-import { ResolvedUserConfig } from "../types.js";
+import { Command, ResolvedUserConfig } from "../types.js";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import { PKG_JSON_PATH } from "../constants.js";
 import { Core } from "../core/index.js";
@@ -12,12 +12,14 @@ import { parseApi } from "../utils/api.js";
 import { compileSync } from "@mdx-js/mdx";
 const pkg = fsx.readJSONSync(PKG_JSON_PATH);
 
-export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
-  const provider = await virtualProvider(config);
-  const instance = Core.getInstance(config);
-  const { sidebars, routes, appData, sandboxes } = await instance.prepare();
+export const docit = async (
+  command: Command,
+  config: ResolvedUserConfig
+): Promise<PluginOption> => {
+  const instance = Core.getInstance(command, config);
+  const virtualPlugins = await instance.prepare();
 
-  const plugin: Plugin = {
+  const docitPlugin: PluginOption = {
     name: "vite-plugin-docit",
     enforce: "pre",
     config: () => {
@@ -26,17 +28,14 @@ export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
         base: config.publicPath,
         optimizeDeps: {
           include: [
+            "@mdx-js/react",
+            "qrcode",
+            "styled-components",
+            "highlight.js",
+            "lodash-es",
             "react",
             "react-dom",
-            "styled-components",
-            "shallowequal",
-            "hoist-non-react-statics",
-            "react-is",
-            "lodash-es",
-            "highlight.js",
-            "qrcode",
           ],
-          exclude: ["@mdx-js/react"],
         },
         build: {
           outDir: path.resolve(process.cwd(), "./docs-dist"),
@@ -53,7 +52,8 @@ export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
         clearScreen: false,
         publicDir: path.resolve(config.docs, "./assets"),
       });
-      return mergeConfig(baseConfig, config.vite);
+      const mergedConfig = mergeConfig(baseConfig, config.vite);
+      return mergedConfig;
     },
     transform(_, id) {
       if (id.endsWith("?needParse")) {
@@ -66,29 +66,31 @@ export const docit = async (config: ResolvedUserConfig): Promise<Plugin[]> => {
       const result = reg.exec(id);
 
       if (result) {
-        const content = instance.getSandBoxMapper().get(result.input);
+        const content = Core.getInstance().getSandBoxMapper().get(result.input);
         const compiled = compileSync(content, getCompilerOptions(config));
         return compiled.value as string;
       }
     },
   };
 
+  const mdxPlugin = await mdx(config);
+  const nodeResolvePlugin = nodeResolve({
+    moduleDirectories: [
+      "node_modules",
+      `./node_modules/${pkg.name}/node_modules`,
+    ],
+  }) as PluginOption;
+  const providerPlugin = virtualProvider(config);
+  const reactPlugin = react({
+    jsxRuntime: "classic",
+  });
+
   return [
-    plugin,
-    appData,
-    routes,
-    sidebars,
-    provider,
-    sandboxes,
-    ...(await mdx(config)),
-    ...(react({
-      jsxRuntime: "classic",
-    }) as Plugin[]),
-    nodeResolve({
-      moduleDirectories: [
-        "node_modules",
-        `./node_modules/${pkg.name}/node_modules`,
-      ],
-    }),
+    docitPlugin,
+    virtualPlugins,
+    providerPlugin,
+    reactPlugin,
+    mdxPlugin,
+    nodeResolvePlugin,
   ];
 };