+++ src/node/plugins/mdx/api.ts
@@ -1,12 +1,14 @@
+import resolve from "resolve";
+import path from "path";
+import { parseMdxToTree } from "../../utils/index.js";
+import { select } from "unist-util-select";
 import { visit } from "unist-util-visit";
 import { Plugin } from "unified";
 import { Link, Text } from "mdast";
-import { ResolvedUserConfig } from "../../types.js";
-import { parseMdxToTree } from "../../utils/index.js";
-import { select } from "unist-util-select";
+import { ResolvedUserConfig } from "../../types";
+
 import { VFile } from "vfile";
-import resolve from "resolve";
-import path from "path";
+
 const { sync } = resolve;
 
 export const api = (config: ResolvedUserConfig): Plugin => {